Introduce separate struct for program and task state
So far we have managed the most of the application's state through a
single object combining both the program state as well as the task
state. We would like to move to a model where those two states are
treated differently, because they are fundamentally different: while the
task state is being mutated by the application, the program state is
effectively reflected through the UI. As such, the latter should be
decomposed and assimilated in the UI and not mutated on the fly like the
task state.
With this change we make this split more pronounced by creating two
first class structs: TaskState represents the task state and ProgState
represents the program (or UI) state. We still have a State struct but
it is mainly used to express the dependency connecting the two during
creation. Objects of this type are intended to be destructed into their
constituent parts.