Rename ProgState to UiState
The concept of "program state" was found to be named not optimally, as
program state conceptually also comprises the task state. So for an
outside person the distinction may not be clear.
To alleviate this problem, this change proposes a different wording: UI
state. In line with that it replaces the usage of ProgState (and
similar) with UiState.