Reorder TaskState and UiState usages
The renaming of program state with UI state had another positive side
effect: the lexicographic ordering of the two terms is changed. As we
would very much like to express the dependency that the UI state has on
the task state by having a consistent ordering in terms of usage of the
two states throughout the program, this change in ordering is very much
welcome.
With this patch we adjust all clients of the two to deal with the task
state before the UI state.