Avoid a couple unnecessary allocations
With the Rust Edition 2018 we have support for non-lexical lifetimes.
With those, the borrow checker is better able to infer accurate
lifetimes of bindings and does not only simply look at the containing
scope.
Because of that added intelligence we are able to save a few allocations
there were necessary beforehand due to conflicting lifetimes.