Demote more lints to "warn"
In the past we have removed the "warnings" lint from the deny group, as
future changes of Rust may warn on more occasions (the usage of
deprecated items being a popular one).
However, there are more lints that may change in the future. It is not
known whether an exact list exists and so with this change we use best
judgment as to what lints to demote to "warn".