Address a couple of clippy suggestions and suppress others
Now that clippy is finally shipping as part of the Rust distribution we
use, we plan to use it more frequently. With this change we address the
minor issues it pointed out and suppress a couple of others that appear
to be pretty subjective (potentially in conflict with a certain code
"style") or otherwise not very helpful.