Pass State object to TermUi constructor
This change continues the effort of refactoring state management with
the eventual goal of persisting more program related state (and loading
it upon startup). More specifically, this patch simplifies the way we
pass the state from the main function to the UI, by using a combined
State object as opposed to splitting it into the constituent parts.