Switch to using the system allocator
In the past we have already taken a couple of steps to reduce the size
of the final binary, arguing that binary size is the metric of most
relevance for the program at hand.
With that in mind, this change decreases the binary size further by
swapping the default allocator we use (typically jemalloc) with the
system allocator (which is malloc based on Unix systems). Given that we
are not allocation sensitive, there is no point in wasting binary size
on something that adds no value.
This change decreases the binary size by another 324 KiB (for an already
stripped release mode binary).