Detect temp directory dynamically
We have hand rolled most of the temporary file logic ourselves, and part
of that means that we need to name a directory to use for the temporary
file. So far we have unconditionally assumed /tmp, but that is not a
very portable choice.
With this change we use std::env::temp_dir to deduce this directory
instead.