Rename Selection* to Iteration*
The selection module came into being as part of the implementation of
search support which would find a certain term on a given tab and select
the tab (and task) containing it. The module itself, however, is a bit
of a misnomer, as it does not really deal with selection per se. Rather,
it is really only supporting the iteration over the various tabs in an
event driven manner.
To make this fact more explicit, this change renames the module and its
member accordingly.