Make UiState a publicly deconstructible wrapper
As it stands the UiState struct is an actual "class" with internal
state. We are moving towards a scheme where the UI's state is
distributed over its constituent parts, and not centralized in a single
object. In such a design the UiState is a mere container used for
grouping and passing around related state, but there is no reason for
secrecy or anything of that sort -- the members should be freely
accessible.
With this patch we make the UiState's members public to allow for easy
deconstruction.