Gather UI state from individual widgets
This change completely reworks how we handle the UI state, the most
important objective being that we now gather all the state on demand in
order to persist it to a file, instead of keeping it in a centralized
location to being with.
In general that makes the design more natural (no more cloning of
queries; only a single source of truth for them) and it was determined
that this refactoring is a requirement to persist more parts of the UI
(such as the currently active tab as well as the selected task), but
gathering all this data is not the most straight forward process in our
componentized system.
Right now (in order to preserve the status quo in terms of what is being
persisted) we mainly have to gather the queries used in the individual
tabs. In order to accomplish that, we use an event together with the
iteration logic we already employ for the search feature.