Update gui crate to 0.3
This change bumps the gui crate version we consume to 0.3. The main
change in this new version (and contributor to most of the changes made
by this patch) is that widgets as defined by this crate no longer work
with a predefined event type. Instead, they are fully generic over the
event type to use.
For this crate, this means that we now define our own event type, which
just wraps a termion key. That in turn means that we can remove the
translation logic which we used to convert a termion event into a gui
specific one.
Furthermore, the signatures of a few methods changed in incompatible
ways, the users of which we adjust accordingly.