Keep raw key event data around
As it stands the readline support is somewhat fudged on top of the other
existing functionality. One manifestation of this lack of integration is
that we work with already parsed keys everywhere throughout the program.
Because libreadline requires the raw key input to work properly (which
is dependent on the terminal emulator), we would effectively have to
reverse the conversion to Key objects (from data as read from the input
stream) as performed by termion. Naturally, that is not something we
want to do.
To achieve proper integration nevertheless, this change instead takes
care of storing the raw input data along with the key object. This is
only a first step. A next step will have to adjust the UI code to work
with this raw data.