Add support for creating a core dump on panic
Debugging panics in the program has always been a painful experience,
mostly due to our use of the terminal's "alternate" screen, which
effectively makes panic messages (which will go to it) disappear in
nirvana after program termination.
While there probably are a bunch of ways to solve this problem, an
additional complication may be that by now reasoning about the program
state just by virtue of what a backtrace provides can be very
challenging.
To that end, this change goes down the route of adding support for
creating a core dump when a panic is encountered to the program. In
order to make this work we have to revert the change to abort on panic
and also create a binary with debug information.
Lastly, this functionality is guarded by a newly introduced feature,
'coredump'. Unfortunately, there seems to be no way to make a profile
configuration dependent on a feature and so we have to live with debug
information being container in the binary irrespective of whether the
feature is enabled or not.
Note that debug information can still easily be stripped off through
strip(1) -- which is something that arguably should be done anyway -- so
the file size increase induced by this change is should not make its way
into the final program.