Remove 'id' member from TabState struct
The TabState struct was used to implement the CollectState event, which
gathers the state of all tabs/queries. The way the result was conveyed
back to the caller was for the caller to pass in its ID, such that once
everything has been collected an event could be delivered directly to
the requestor.
Such a directed event is unnecessary, however, when the requestor is
known to be a parent of the TabBar, as is always the case here. In fact
requiring such an ID makes it harder to harness this functionality for
testing, which is something we want to do in the future.
To that end, this change simply removes the ID from the TabState struct.