Move 'selected' query field into TabState struct
The 'CollectedState' event contains two elements, a vector of queries
and the index of the currently selected one.
Given that the TabState struct already contains a vector of queries as
well, this change moves the selection index in there. This step allows
us to just pass through TabState objects, which will be helpful in the
future when we want those to carry additional information that is just
passed around without intermediate handlers of the event having to care
about it.