Collect code coverage as part of CI pipeline
With this change we collect code coverage information as part of the
GitLab CI pipeline. We use kcov which is able to gather such metrics
from a binary containing debug symbols, i.e., without recompilation to
provide additional instrumentation information.
The coverage report (in the form of web page detailing individual files
and lines of code) is included as an artifact to the pipeline run and
can be downloaded on demand. Unfortunately, it is not completely
straight forward to access it. In general it can be accessed at:

  https://d-e-s-o.gitlab.io/-/notnow/-/jobs/<job-id>/artifacts/kcov/kcov-merged/index.html