Fix assertion failure when searching after aborting search term entry
When we start entering a term to search for but abort this entry for
whatever reason, the next attempt to search will panic. The problem here
is that we never anticipated to be in the "Preparing" phase when
starting a search and so we asserted that it would not happen. This
change corrects this misunderstanding by handling this case more
gracefully.