Lowercase Pos and Neg fields of serialized queries
Query objects, when serialized, are the only objects that contain fields
that start with an uppercase character, namely 'Pos' and 'Neg'. All
other objects are solely comprised of lowercase letters.
In an attempt to achieve greater consistency, this change makes sure
that we serialize those two fields using lowercased names. For the sake
of not breaking existing configurations needlessly, we make sure to
still recognize the uppercased variants as well.