Rework multi-byte character support
A change made a while back essentially caused us to ignore any
multi-byte character input. That step was necessary because with the
existing logic we were potentially removing bytes from a multi-byte
character, which could invalidate the character altogether, invalidating
invariants by the String struct.
This change revisits our approach to this problem. Instead of silently
dropping input of characters, we allow their input and make character
insertion/removal and cursor navigation honor character boundaries
properly.
Note that this approach will still not work correctly with all Unicode
characters (an example being 'y̆'), in the sense that those may still be
displayed wrongly (with the cursor at the wrong location). However, all
String imposed invariants will be upheld correctly.
In order to make those characters work we would need to use the
unicode-separation crate, which would allow us to work on grapheme
clusters instead of characters (the aforementioned 'y̆' is comprised of
two characters but a single grapheme cluster).