Run clippy without readline feature enabled as well
So far we have run clippy with the --all-features switch set. As the
name implies, this parameter causes the linting to happen with all
features enabled. The readline feature, that the program supports,
however, isn't really additive but rather enables a couple of different
code paths. As such, it makes sense to run the linter with and without
this feature enabled.
To that end, this change adjusts the GitLab CI pipeline configuration to
run clippy with both configurations.