Use locked stdout for output
Writes to the process' standard output are synchronized within the
program through a shared lock. This lock is acquired each and every time
a write is performed. We only have a single point of usage of the stdout
handle and so all such synchronization wasted effort.
Hence, with this change we lock the handle once and work with the locked
instance.