Simplify TermUi::handle_custom_event method
In earlier version of Rust it was not possible to use pattern guards on
variants that would move an attribute, making the logic inside the
TermUi::handle_custom_event method more complicated than it had to be.
The corresponding feature got stabilized with Rust 1.39 and because we
require this version by now, this change simplifies the logic inside
this method, making use of a pattern guard in the match expression.