Rename intra_doc_link_resolution_failure lint
The intra_doc_link_resolution_failure lint has been renamed to
broken_intra_doc_links in recent versions of Rust. With this change we
update it accordingly in our list of lints to use for the crate. Because
our CI tests against the minimum supported Rust version, which does not
yet understand the new lint, we configure it to allow unknown lints.