Rename TermUiEvent to Message
With an upcoming release the gui crate will move to a different event
design: instead of clubbing "regular" events and custom events together
they will stay separate and have different purposes. In fact, events in
the new version's sense can no longer be emitted at will from within the
UI itself. Rather, this capability will be provided by messages.
In order to prepare for this change in terminology, this change renames
the TermUiEvent event type to Message. This type has been used as a
custom event type so far and as such will directly map to messages in
the new world order.