Update gui crate to 0.6.0-alpha.0
The gui crate is undergoing a major and multi-step transformation as
part of which it will support async event handling. With this change we
integrate with the first alpha release of version 0.6 of the crate,
which marks a first step on this journey. As part of this release
widgets are separated from their mutable data, with an explicit
retrieval being necessary.