Use send logic for retrieving test state
In order to do their job, our tests need to peek into the UI's state to
evaluate results. So far we have done so by sending a custom event and
evaluating the response.
With version 0.6.0 of the gui crate, events are no longer intended for
this purpose. Rather, messages can be used in their stead.
To that end, this change converts the test code as well as the UI
internal handler logic over to using messages for this state retrieval.