Rework state collection
We have a bunch of logic dedicated to what we refer to as "state
collection" -- the act of retrieving all queries used in the different
tabs as well as their selection states. We require this state in both
the test as well as as part of the saving procedure (which persists the
queries).
This logic has traditionally been rather cumbersome, because it spanned
multiple widgets, which required us to use message passing for
collecting this state. At the same time, we couldn't really keep the
accumulated state around locally, and so we ended up passing it around.
The end result was a working but not particularly easy to understand bit
of communication code.
Luckily, with the gui crate supporting async/await, we are now actually
able to write all this collection code in a simple iterative manner,
making it much easier to follow.
This change migrates the state collection code over to using the new
send/call functionality, which integrates with async/await.