Handle SetInOut event solely through message passing
With this change we adjust all clients using the Message::SetInOut
variant to use the new message passing approach over the custom events
used earlier. In so doing we make the communication end points more
explicit. Furthermore, to make this directed communication still more to
the point, we also pass the widget ID of the In/Out area to the TabBar
and the various TaskListBox objects.