Convert search functionality
This change converts the search from using a decentralized approach to a
centralized one. To that end, we remove the usage of the IterationState
object from the SearchTask message. While at it, we add two more tests
for the search functionality, one on an empty tab and the other on one
with a single task only.