Remove distributed iteration logic
When we originally devised the interaction between the widgets making up
the UI we had to use a fully distributed approach, because no single
entity could coordinate operations such as task selection or task
search. The main reason for this restriction was our event based
communication scheme, which required each widget to return the events it
wanted handled.
Now that we have async event handling available, we can revisit this
decision. With this patch specifically we adjust the task selection
logic, which is the last component using the distributed iteration
logic, to be coordinated by the TabBar.