Remove remaining non-hook driven custom event handling
This change moves the last non-hook driven custom event handling
infrastructure to a message based model. With this step we only have a
single custom event remaining, which is used by the event hook that we
install. Removal of this custom event will only be possible after
continued changes in the gui crate itself.