Update gui crate to 0.6.0-alpha.2
With this change we integrate with version 0.6.0-alpha.2 of the gui
crate. In this version the hook logic has been reworked. For this crate,
this means we now have to provide a way for events to be merged.
Furthermore we switch from sending ourselves an event from the event
hook (sort of as a callback mechanism that triggers after the hooked
event has been handled) to using the newly introduced post-hook
functionality.