Create tasks after "cloned" ones
When we create a new task we always place it at the end of the task list
and with that at the end of the displayed query. That can be a rather
unsatisfying behavior when task lists are long and the intention is to
add a new task after an existing one.
To make the behavior a bit more user friendly, with this change, when we
create a new task we make sure to place it right after the "cloned" one.