Make TaskListBoxData::selected_task return an Option
The TaskListBoxData::selected_task method is fairly confusing to use and
not very "rusty": it requires that the query we work on has tasks and
will panic otherwise, which in turns means an additional check
beforehand that is easy to forget. That's not in line with how most Rust
APIs behave.
This change fixes the problem by making the function returning an Option
instead of unwrapping internally.