Remove unused_result lint from warn list
We were arguably a bit too liberal with enabling warnings in the
program. Specifically, warning unconditionally on unused results turned
out to be more of nuisance than helpful.
To that end, this change removes the 'unused_results' lint from our list
of warn lints, falling back to the default behavior.