Format the source code using rustfmt
This change applies changes as suggested by rustfmt to the code base. We
were hoping that over time we could just incrementally fix problems as
we were touching the respective functions, but for the majority of the
tests as well as the event handlers this is simply unrealistic. At the
same time, not having proper formatting for existing code means that we
pay a toll for newly introduced code as well, because violations are
much harder to catch as they drown in all the suggestions.
To that end, this change takes the last resort approach and formats
everything.