Factor out Modal trait
Our different widgets tend to contain similar or at times even identical
functionality. In such cases, it makes sense to try and identify common
abstractions and factor them out into isolated and reusable components.
In the future we will have another widget that needs to be conditionally
focused for a certain time, and then later transfer focus back to the
originally focused widget.
In an attempt to identify a proper abstraction and to make this
functionality reusable, this change factors out a new trait, Modal, that
encapsulates the logic of remembering focused widgets and focusing as
appropriate.