Rename TaskListBox selection functions
This change renames the TaskListBox's functions that allow for changing
the currently selected task. Specifically, we rename select to
change_selection and set_select to select. This step is mostly taken
because this naming seems more explicit, but we will also factor out
the selection bits with an upcoming change, and these are the names we
will use.