Add logic for rendering the dialog
This change adds the necessary logic for rendering the dialog widget.
The dialog is displayed in the center of the terminal and only covering
part of the available space.
Note that contrary to other widgets, because of the desired notion of
creating an overlay effect, we have to make sure that the dialog renders
something to each cell that it covers -- it cannot just leave some cells
untouched, because that would partly destroy the overlay effect. It is
for that matter that the rendering code is structured differently
compared to that for, say, the task list box. In the future, we hope to
transition all widgets to such a rendering approach, and to ultimately
remove the full screen clear that we do as a pre-rendering step.