Hook up dialog to 't' key
With this change we hook up the dialog from the task list box. By
pressing 't' we show and focus it, allowing us to modify a task's tags.
Using Esc/'q' the operation can be aborted, or return confirms the
changes and updates the underlying task's tags according to the
selections made. Along with that enablement we add a bunch of tests.