Add support for quickly jumping to tags
This change adds support for quickly jumping to a tag on the dialog.
Specifically, we use the 'f' and 'F' keys to get into forward or
backward "jump to" mode, respectively. If subsequently the user presses
a key that a tag below (forward) or above (backward) starts with, that
tag will be made the selected one. We may eventually introduce similar
functionality for tasks, but given that tags are sorted they seem like
a more suitable candidate.