Switch to using tarpaulin to measure code coverage
So far we have used kcov to collect code coverage information for the
project. It has repeatedly been broken in the past, for various reasons.
It also is a pain to set up locally and the report is arguably not very
trustworthy.
With this change we switch over to using tarpaulin to measure code
coverage for the project. Like kcov it can generate an HTML report
(although not quite as nice) and it is more precise in its measurement.
Lastly, setting it up locally is much less of a hassle.