Make task addition and move one "atomic" unit
With upcoming changes we plan to add support for undoing (and redoing)
operations performed on tasks. For that to work, we will represent
certain "high level" task operations as objects and keep track of them.
Unfortunately, right now such tracking would not be entirely as
intended, though, because the task list box, when adding a new task,
actually performs two such high level operations: an addition and move
to the expected position. That is a problem, because it would mean that
both operations could be undone separately, which is bound to cause
confusion.
To fix this issue, this change adjusts the Tasks::add method to
operationally accept the ID of the task after which the new task will be
placed.