Introduce find_idx function
This change introduces a new utility function to the task management
layer: find_idx. As the name implies, the function can be used to find
the index of a task given its ID.
The purpose of this function is two-fold: First, it allows us to shorten
the code and, arguably simplifies the logic as well, as is demonstrated
by the change in addition to the mere function addition. Second, and
more importantly, it concentrates a lot the module's "unwraps" into a
single function. That's helpful, because the assumption that we can
always find a task with a provided ID is at the core of the module.