Introduce infrastructure for tracking task (and other) operations
With upcoming changes we would like to introduce undo & redo
functionality for actions pertaining tasks to the program. In order to
do that we need some way to stash away a task related operation while
also being able to revert it.
For the storage part, we use a fixed size ring buffer so what we can
provide the user with a bounded number of undo steps.
To model the ability to revert an operation, we introduce the Op trait,
which contains two methods: exec and undo. Implementors should ensure
that undo reverts all actions performed by exec.