Introduce task specific operation support
With a recent change we gained the ability to model operations and their
inverse in a generic way. With this change we build on top of this
functionality to bring task specific operations into this framework.
Specifically, we introduce a new enumeration type, TaskOp, that can
represent any supported operation on a set of tasks, while also
providing a way to revert said operation.