Reuse add_task in more places
We have three locations in the tasks module that logically are concerned
with adding a task to a vector of tasks. However, only two use the
add_task function.
With this change we adjust add_task in such a way that it can cater to
the needs of all three clients, making the code slightly easier to
understand and reducing duplication.