Keep track of executed operations in Tasks object
This change embeds an object of our recently introduced Ops type into
each Tasks object and switches all task related operations over to using
TaskOp objects. That means that we can now undo and redo task related
operations (although access to this functionality is not yet plumbed
through the UI to be actually usable by users).
As it stands we remember up to 64 past operations, though this maximum
can easily be bumped up (and potentially be made user-configurable in
the future).