Introduce undo and redo support
This change builds on the functionality introduced in earlier changes to
provide a way for users to undo the last task related operation as well
as redo it. The ability to do so has been a long desired feature,
because the repercussions of an accidental wrong key press have the
potential to other destroy relevant data and the only recourse so far
has been to quite the program without saving. That, in turn, can a
source of data loss as well, though, if the user is unsure of the last
time data was saved.
Note that while the core feature is functionally complete, the UI for it
has still room for improvement. Specifically, the current task selection
is preserved, which makes it non-obvious where a change occurred.