Allow for task operations to return a value
This change adjusts the Op trait to allow the exec and undo methods to
return a value. In the future, we intend to use this capability to have
them return some kind of identifier of the task that the operation
pertained. That, in turn, will allow us to properly select the task
being worked on as part of the undo/redo workflow, which is essential to
improve the user experience.