Properly select target of undone/redone task operation
The recently introduced undo/redo functionality had one glaring
usability issue: the task being target by the operation was never
actually selected, so it was very hard for a user to tell what has
changed as part of it.
This change fixes this issue by bubbling up the index of the task the
most recent operation worked on, translating this index into a task ID,
and then selecting the task with this ID via our regular task selection
logic.
Note that there is still a rather large deficiency in this logic: it
does not work on task removals. That is, the selection will not be
changed when a task addition is undone or when a task removal is redone.
There are some arguably insurmountable problems with finding the correct
task to select in these cases. The reason is that we are working at a
layer where we have access to the task database, but don't have any
knowledge of a particular view on it. The view defines which task is
displayed before any other and without its involvement we cannot know
what the task before or after a removed on is -- thus, we cannot select
such a task either.