Use latest stable Rust image for code coverage
When we introduced the tarpaulin code coverage infrastructure, we bumped
the minimum supported Rust version to accommodate for the requirements
imposed by cargo-tarpaulin. However, we can actually just decouple the
tarpaulin jobs from the "regular" build jobs and use a different image
there.
With this change we go down that road, simply using the most recent
stable Rust image.