Measure code coverage on all branches
As it turns out, GitLab actually provides a setting that controls
whether or not to have automatic expiration of the artifacts of the most
recent (successful) pipeline run. We now have this setting enabled and
so with this change we remove the explicit expiration setting we had in
the CI configuration so far. Because of that change, we also no longer
have a good reason for gathering code coverage only on the master
branch.