Re-select task after tag change
When we toggle the completion flag using space (' '), we find and
re-select the task on whatever query it is located now (if any). When we
introduced the dialog based tag editing feature, on the other hand, we
did not add such a selection step. That is a confusing behavior.
To fix this shortcoming, this change makes sure that we re-select the
task after any of its tags were changed.