Add support for searching for an exact task summary
So far our task search was always a (case-insensitive) sub-string
search. That is, a task was "found" if the string to search for was
contained in its summary.
With upcoming changes we would also like to support searching for an
exact task summary. That is, one that is a literal (and case-sensitive)
character for character match.
This change introduces the capability for such a search. Right now it is
not yet reachable through the UI and effectively dead code.