Add support for finding the current task on other tabs
Depending on a user's workflow and setup, a task can reasonably appear
on multiple tabs, because tabs may provide overlapping views on the
underlying data. Sometimes it is useful to find the very task at hand on
all remaining tabs.
This change adds support for searching for all tabs that the currently
selected task appears on via the '*' key (the idea of which we borrowed
from Vim). The feature piggy-backs on the existing task search, but does
away with the case-insensitivity and sub-string matching and instead
only matches tasks with the same summary. Note that strictly speaking
different tasks could have the same summary and we would select them as
part of the search if that is the case. We argue that in most cases that
is what the user wants anyway and it feels like the expected behavior
given that the UI makes it clear that a regular task search is going on.