Allow deref_into_dyn_supertrait lint
We have created our own version of trait upcasting by implementing Deref
to Cap for MutCap. Because the compiler automatically invokes deref we
effectively can use a &MutCap where a &Cap is required.
It appears, however, as if Rust may support trait upcasting at some
point in the future. For now, they have introduced a lint that flags
implicit upcasting to a super trait via Deref, in an attempt to force
developers to check that their usage of the pattern is in line with the
functionality that the unstable trait_upcasting feature provides. We
believe that is the case for this crate and so we now allow the lint.