Shorten color serialization and deserialization code
The deserialization for our Color type is basically the serde proc macro
generated output, copied in place. As such it is very verbose and not
friendly to readers.
With this change we replace this complex logic with four functions that
take care of serializing and deserializing the Reset and Rgb variants,
respectively.