Make Tag representation transparent
This change makes our serialization format of a tag more compact by
declaring the type as having a transparent serde representation. As a
result, a tag is no longer serialized as a structured type, but merely
as a number.

To convert notnow.json and tasks.json to adhere to the changed format,
the following Vim command may be used:
  :%s!{\n[ ]\+"id": \([0-9]\+\)\n[ ]\+}!\1!g