Rename "query" related colors to "tab"
The usage of the term "query" throughout the program is arguably not the
most apt. That's even more so as a query is meant to be an internal
concept. However, currently it is leaked into the user-facing
configuration, where we have color names that have "query" in their
name.
With this change we rename these colors to use the term "tab" instead,
which should make it clearer to users what they represent.

To convert the notnow.json configuration file from Vim, the following
command can be used:
  :%s!_query_\([fb]g\)"!_tab_\1"!g