Rename "query" to "view" in the ser module
We have decided to change nomenclature within the crate a bit.
Specifically, the term "query", which we used quite pervasively, is not
the best in conveying the underlying concept. We have found "view" to be
more representative.
With this change we start the rename by adjusting the ser module, which
contains our serialization and deserialization types, to use "view"
instead of "query". Subsequent changes will adjust the remaining usages
of the term.

To convert the notnow.json configuration file to use the new term, the
following command may be used:
  :%s!"queries": \[!"views": \[!g