Adjust misleading comment in search_tasks_on_multiple_tabs test
The search_tasks_on_multiple_tabs has a comment in it stating that we
activate the third tab. However, the key pressed being fed in actually
just navigate to the second one. It so happens that in both cases the
outcome of the key sequence is exactly the same, meaning that the test
succeeds in either case. The comment is still misleading, though. With
this change we adjust it to refer to the second tab instead.