Fix reverse search when on the first task on a tab
As it turns out, when searching backwards and landing on the first
task on tab, any attempts to search further back from there would not
result in any action being taken. The reason for this behavior is that
we were using saturating subtractions in the reverse find logic and so
when we went "before" the first element, we would just stay on it
instead. That effectively caused us to never go before this one task.

This change fixes the issue by using checked subtraction (together with
early returns) to handle this case correctly.