Move most of main.rs into new lib.rs
With upcoming changes we would like to use some of the application start
up code in an example binary. Currently this logic is all contained in
main.rs, from which it cannot be imported.
With this change we introduce a library portion to the crate in a newly
added module, lib.rs, and move the majority of this start up code over
there.