Add example program showing notnow UI with test data
This change introduces an example program that illustrates notnow's UI
with some of the example data we generate for testing purposes. This
example program has two use cases. For one, it illustrates the program's
UI without the need for a user to do anything. And second, given that it
uses the same data that we also use in many of the tests, it is now much
easier to simply "replay" test events in this example than to think
through the contents of the various tabs. This purely mental process has
repeatedly been shown to be complicated and error prone, making it
hard to develop tests and more likely to introduce bugs in them. In
order to power this example, we also have to export some more of the
crate's functionality publicly. Of course we do not intend to publish
the crate as a library, and so really this functionality is still not
available to arbitrary users.