Make Template type cheaply shareable
Basically by design, our Template type has to be shared: it constitutes
the "template" from which tags can be instantiated and each tag has a
back-reference to the underlying template.
So far we have expressed this fact explicitly, by wrapping each Template
object in an Rc. This approach is very verbose and arguably leaks
what should be implementation details to clients.
With this change we move the usage of Rc internally to the type, making
it directly cheaply clonable. Clients are adjusted to no longer use Rc
on their end.