Use tempfile crate as provider of temporary file functionality
In the early beginnings of the program we introduced our own temporary
file functionality, in order to cut down on external dependencies. With
upcoming changes we will also require support for the creation of
temporary directories, the support for which we have no interest in
rolling ourselves.
Given this new requirement and the fact that we need this functionality
only for testing purposes, this change replaces our ghetto NamedTempFile
implementation with the type of the same name from the newly introduced
dev-dependency to the tempfile crate.