Rename {load,save}_state functions
With upcoming changes we would like to make further changes to the way
we store tasks. Specifically, instead of having everything stored in a
single file we are aiming for a scheme where each task is stored in one
file.
Internally, that will mean we need an additional way to serialize (and
deserialize) tasks -- not just from a file but also from a folder
(containing many files). In preparation for that, this change renames
the load_state and save_state functions, by appending the _from_file
prefix.