Use a consistent ordering of UI and task state in code
We have a bit of an inconsistency in the code when it comes to the
ordering of UI state versus task state. Sometimes the former is dealt
with before the latter and sometimes vice versa.
With this change we adjust all locations to make sure to always
reference the UI state before task state.