Unify naming of configuration file variables
Currently we use the term 'ui_path' (or 'task_path') for variables
holding the path to the UI configuration file. That may not be the best
choice, because these variables are already of type Path or PathBuf, but
it is also not done completely consistently, as we do have occurrences
of 'ui_config' for the same purpose as well.
With this change we try and standardize the usage of 'ui_config' (and
'task_config') for these variables.