Introduce and use ser::Task::with_tags setter
For testing purposes only, we have a constructor for the ser::Task type
that expects nothing but a summary. That turned out to be fairly
convenient, but it excluded all callers that require tags on the task in
question.
To mitigate this issue, this change introduces a new setter, with_tags,
to the type, that can be used for conveniently setting the tags shortly
after object construction.