Introduce task database
Over time it has become more and more clear that associating a unique ID
with a task the way we do is not really a great solution. The main
problem is with our approach is that global uniqueness -- as we require
it -- is hard to ensure and unnecessarily strong a requirement. That has
resulted in our task IDs solely being in-memory entities, never to be
persisted. That in turn has trickle-down effects such as us not being
able to persist our undo stack, because it relies on such IDs.

This change marks the first step in mitigating this issue. It introduces
a task database, an instance of which will hold the program's tasks in
the future. In this model it is the database that assigns IDs to objects
in it -- tasks --, but tasks themselves no longer carry an ID. In so
doing, we push the uniqueness requirement from the global level down to
the individual database instance. A database can, conceptually, be
constructed with passed-in IDs while checking uniqueness, but it may
also assign IDs as objects are inserted. As such, we gain much more
flexibility in working with said identifiers. Furthermore, we simplify
working with tasks themselves, as they can be constructed and worked
with outside a database context where doing so is sufficient and we no
longer have to worry about uniqueness constraints in any way.

Along with the database we also introduce the Idable trait. This trait
captures the ability of objects to report a unique identifier and will
be used until we eventually can migrate away from using task IDs to
database IDs. As such, this trait is a temporary means allowing us to
migrate more and more parts of the program to using the database over
time, until we can eventually get rid of the concept of a task ID (and
this trait) altogether.